# openapi.yaml
openapi: 3.1.0
info:
  title: Ключи API
  version: 0.3.6
  description: 'Реализованный API «Ключи» 0.3.6. Квартира/комнаты, группы арендаторов, настраиваемые счётчики и
    регулярные начисления. Payload и переходы: docs/API.md. Суммы в копейках, показания в тысячных. Перевод денег
    не выполняется.'
servers:
- url: /
  description: Same origin
security:
- BearerAuth: []
- CookieSession: []
paths:
...
  /api/config:
    get:
      summary: config — публичные настройки, без секретов
      operationId: config
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /api/auth/demo:
    post:
      summary: demoLogin — только в demo/test
      operationId: demoLogin
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
      parameters:
      - in: header
        name: X-Keys-Client
        required: true
        schema:
          type: string
          enum:
          - miniapp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - role
              properties:
                role:
                  enum:
                  - tenant
                  - owner
  /api/auth/max:
    post:
      summary: maxLogin — проверка подписанной строки MAX
      operationId: maxLogin
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
      parameters:
      - in: header
        name: X-Keys-Client
        required: true
        schema:
          type: string
          enum:
          - miniapp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - initData
              properties:
                initData:
                  type: string
                  maxLength: 16384
  /api/auth/logout:
    post:
      summary: logout — завершить текущий сеанс
      operationId: logout
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - in: header
        name: X-Keys-Client
        required: true
        schema:
          type: string
          enum:
          - miniapp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
  /api/auth/revoke:
    post:
      summary: revokeSessions — завершить все сеансы
      operationId: revokeSessions
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка запроса/доступа/состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - in: header
        name: X-Keys-Client
        required: true
        schema:
          type: string
          enum:
          - miniapp
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    CookieSession:
      type: apiKey
      in: cookie
      name: keys_session
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Max-Bot-Api-Secret
  schemas:
    Error:
      type: object
      required:
      - error
      - requestId
      properties:
        error:
          type: string
        requestId:
          type: string
...